changed precision to double!

+++ include/ppt/types.hpp
@@ -9,7 +9,7 @@
 #include "math.h"
 
 // SELECT FLOATING-POINT PRECISION (DATA-TYPE)
-using float_type = float;
+using float_type = double;
 
 // SELECT EXECUTION-SPACE BASED ON COMPILE-TIME CHOICE
 #if defined(PPT_ENABLE_CUDA_BACKEND)
